requirements.txt
kafka-python==2.0.2
six
pyspark==3.4.1
